rename: /login_page/main.js -> Login.js
