+++ src/App.js
@@ -13,7 +13,7 @@ import ScrollToTop from "layouts/ScrollToTop";
 import Main from "./pages/main_page/Main";
 import Apply from "./pages/apply_page/Apply";
 import Introduce from "./pages/doing_page/Introduce";
-import { Login } from "./pages/login_page/Main";
+import { Login } from "./pages/login_page/Login";
 import ApplyGuide from "./pages/apply_page/applyGuide/ApplyGuide";
 import Not_allow_apply from "pages/Not_allow_page/Not_allow_apply";
 import { MiddleResult } from "pages/result_page/MiddleResult";
