+++ src/pages/doing_page/Introduce_5/Introduce_5.js
@@ -1,5 +1,5 @@
 import React from "react";
-import MainApplyCard from "../../main_page/MainApplyCard";
+import MainApplyCard from "../../main_page/IntroduceMain/ApplyCardMain/MainApplyCard";
 import "../Introduce.css";
 import "./Introduce_5.css";
